Ishiiruka Dolphin 5.0 Stable build 744
This is a fork of the official Dolphin Wii and Gamecube emulator
providing reduced graphics cpu usage and better compatibility with old
devices.
I've built the .app bundle from my own fork which is up-to-date with
Tinob/master which is the official repo. They don't provide any Mac
builds so I thought why not provide my own.